refactor(backend): organize source files into modules
