+++ packages/backend/src/index.ts
@@ -1,3 +1,3 @@
-import { CompositionRoot } from './CompositionRoot';
+import { CompositionRoot } from './shared/CompositionRoot';
 
 CompositionRoot.Create().getWebServer().start();
