+++ packages/backend/src/shared/Repository.ts
@@ -1,5 +1,5 @@
 import { PrismaClient } from '@prisma/client';
-import { Database } from '../Database';
+import { Database } from './database/Database';
 
 export abstract class Repository {
   constructor(protected readonly database: Database) {}
