+++ packages/backend/tests/e2e/registration.test.ts
@@ -4,7 +4,7 @@ import path from 'path';
 import supertest from 'supertest';
 import { UserInput } from '@dddforum/shared/src/api/users';
 import { UserInputBuilder } from '@dddforum/shared/tests/support/builders/UserInputBuilder';
-import { CompositionRoot } from '../../src/CompositionRoot';
+import { CompositionRoot } from '../../src/shared/CompositionRoot';
 import { DatabaseFixtures } from '../support/fixtures/DatabaseFixtures';
 import { EmailAlreadyInUseException, UsernameAlreadyTakenException } from '@dddforum/shared/src/errors/exceptions';
 import { ValidationError } from '@dddforum/shared/src/errors/errors';
