+++ packages/backend/tests/support/fixtures/PostBuilder.ts
@@ -1,5 +1,5 @@
 import { Post } from '@prisma/client';
-import { prisma } from '../../../src/prisma';
+import { prisma } from '../../../src/shared/database/prisma/prisma';
 import { faker } from '@faker-js/faker';
 
 export class PostBuilder {
